enhancement: refactor contests calls
